Relocate a test function as suggested by Matthew Roeschke (@mroeschke)
